Bug: Reset button not working
The reset button did not work as expected: the input fields were able to reset after the button was clicked, but the results of the previous calculation still remained. The bug was fixed by simply adding an OnClick Function that resets the Results Container by simply assigning it an empty string.